requirements.txt
asyncio==3.4.3
pymongo[srv]==3.11.3
pymongo==3.11.3
numpy==1.21.0
pandas==1.1.3
buttons==0.1.9
Paginator==0.5.1
emojis==0.6.0
TagScriptEngine==2.3.0
pyparsing==2.4.6
motor==2.3.1
pillow
image==1.5.33
colour==0.1.5
webcolors==1.11.1
discord-webhook==0.13.0
humanfriendly==9.2
nextcord==2.0.0a8